Added login/logout functionality.

+++ listify/src/main/java/listify/controllers/APIController.java
@@ -2,6 +2,9 @@ package listify.controllers;
 
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -11,6 +14,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -34,9 +39,11 @@ public class APIController {
 	}
 	
 	@PostMapping("/API/login")
-	public ResponseEntity<String> getPage(@RequestBody Map<String, String> body){
+	public ResponseEntity<String> getPage(HttpSession session, 
+										  @RequestBody Map<String, String> body){
 		String username = listifyService.login(body.get("email"), body.get("password"));
 		if(username != null) {
+			session.setAttribute("username", username); //set session attribute
 			return ResponseEntity.ok().body(username);
 		}else {
 			return ResponseEntity.notFound().build();
@@ -44,9 +51,22 @@ public class APIController {
 	}
 	
 	@PostMapping("/API/register")
-	public ResponseEntity createUser(@RequestBody Map<String, String> body){
+	public ResponseEntity createUser(HttpSession session, 
+									 @RequestBody Map<String, String> body){
 		if(listifyService.createUser(body.get("email"), body.get("username"), body.get("password"))) {
+			session.setAttribute("username", body.get("username")); //set session attribute
 			return ResponseEntity.status(HttpStatus.CREATED).build(); 
+		}else {
+			return ResponseEntity.notFound().build();
+		}
+	}
+	
+	@PostMapping("/API/{username}/logout")
+	public ResponseEntity logout(HttpSession session){
+		String username = (String) session.getAttribute("username");
+		if(username != null) {
+			session.invalidate(); //destroy the session
+			return ResponseEntity.ok().build(); 
 		}else {
 			return ResponseEntity.notFound().build();
 		}
@@ -54,15 +74,24 @@ public class APIController {
 	
 	@PostMapping("/API/{username}/createList")
 	@ResponseBody
-	public String createList(@PathVariable(value="username") String username, 
+	public String createList(HttpSession session,
+							 @PathVariable(value="username") String username, 
 							 @RequestBody String listName){
+		//check if a logged user is trying to access the data of another user
+		if(!(username.equals(session.getAttribute("username")))) {
+			return "-1";
+		}
 		return "" + listifyService.createList(username, listName);
 	}
 	
 	@PutMapping("/API/{username}/updateListName/{listId}")
-	public ResponseEntity updateList(@PathVariable(value="username") String username, 
+	public ResponseEntity updateList(HttpSession session,
+									 @PathVariable(value="username") String username, 
 									 @PathVariable(value="listId") int listId,
 									 @RequestBody String newListName) {
+		if(!(username.equals(session.getAttribute("username")))) {
+			return ResponseEntity.notFound().build();
+		}
 		if(listifyService.updateToDoListName(username, listId, newListName)) {
 			return ResponseEntity.ok().build(); 
 		}else {
@@ -71,8 +100,12 @@ public class APIController {
 	}
 	
 	@DeleteMapping("/API/{username}/deleteList/{listId}")
-	public ResponseEntity deleteList(@PathVariable(value="username") String username, 
+	public ResponseEntity deleteList(HttpSession session,
+									 @PathVariable(value="username") String username, 
 			 						 @PathVariable(value="listId") int listId){
+		if(!(username.equals(session.getAttribute("username")))) {
+			return ResponseEntity.notFound().build();
+		}
 		if(listifyService.deleteList(username, listId)) {
 			return ResponseEntity.ok().build(); 
 		}else {
@@ -82,17 +115,25 @@ public class APIController {
 	
 	@PostMapping("/API/{username}/toDoList/{listId}/createActivity")
 	@ResponseBody
-	public String createActivity(@PathVariable(value="username") String username, 
+	public String createActivity(HttpSession session,
+								 @PathVariable(value="username") String username, 
 			 					 @PathVariable(value="listId") int listId,
 			 					 @RequestBody Activity activity) {
+		if(!(username.equals(session.getAttribute("username")))) {
+			return "-1";
+		}
 		return "" + listifyService.createActivity(username, listId, activity);
 	}
 	
 	@PutMapping("/API/{username}/updateList/{listId}/updateActivity/{activityId}")
-	public ResponseEntity updateActivity(@PathVariable(value="username") String username, 
+	public ResponseEntity updateActivity(HttpSession session,
+										 @PathVariable(value="username") String username, 
 			 							 @PathVariable(value="listId") int listId,
 			 							 @PathVariable(value="activityId") int activityId,
 			 							 @RequestBody Activity activity) {
+		if(!(username.equals(session.getAttribute("username")))) {
+			return ResponseEntity.notFound().build();
+		}
 		if(listifyService.updateActivity(username, listId, activityId, activity)) {
 			return ResponseEntity.ok().build(); 
 		}else {
@@ -101,10 +142,14 @@ public class APIController {
 	}
 	
 	@PutMapping("/API/{username}/updateList/{listId}/updateActivityCategory/{activityId}")
-	public ResponseEntity updateActivity(@PathVariable(value="username") String username, 
+	public ResponseEntity updateActivity(HttpSession session,
+										 @PathVariable(value="username") String username, 
 			 							 @PathVariable(value="listId") int listId,
 			 							 @PathVariable(value="activityId") int activityId,
 			 							 @RequestBody String category) {
+		if(!(username.equals(session.getAttribute("username")))) {
+			return ResponseEntity.notFound().build();
+		}
 		if(listifyService.updateActivityCategory(username, listId, activityId, category)) {
 			return ResponseEntity.ok().build(); 
 		}else {
@@ -113,9 +158,13 @@ public class APIController {
 	}
 	
 	@DeleteMapping("/API/{username}/updateList/{listId}/deleteActivity/{activityId}")
-	public ResponseEntity deleteActivity(@PathVariable(value="username") String username, 
+	public ResponseEntity deleteActivity(HttpSession session,
+										 @PathVariable(value="username") String username, 
 			 							 @PathVariable(value="listId") int listId,
 			 							 @PathVariable(value="activityId") int activityId) {
+		if(!(username.equals(session.getAttribute("username")))) {
+			return ResponseEntity.notFound().build();
+		}
 		if(listifyService.deleteActivity(username, listId, activityId)) {
 			return ResponseEntity.ok().build(); 
 		}else {
